Insert some explicit casts to squash warnings.
Former-commit-id: ec722a7288030d660cde853ad5aa0de177eb1538

+++ src/runtime/runtime.opencl_host.cpp
@@ -63,7 +63,7 @@ extern "C" {
     float msec;                                             \
     cuEventElapsedTime(&msec, __start, __end);              \
     printf(stderr, "Do %s\n", str);                         \
-    printf("   (took %fms, t=%lld)\n", msec, halide_current_time_ns());  \
+    printf("   (took %fms, t=%lld)\n", msec, (long long)halide_current_time_ns()); \
 } halide_current_time_ns() // just *some* expression fragment after which it's legal to put a ;
 #else
 #define TIME_START()
@@ -155,7 +155,8 @@ WEAK void halide_init_kernels(const char* src) {
         dev = devices[deviceCount-1];
 
         #ifndef NDEBUG
-        halide_printf("Got device %lld, about to create context (t=%lld)\n", (long long)dev, halide_current_time_ns());
+        halide_printf("Got device %lld, about to create context (t=%lld)\n",
+                      (long long)dev, (long long)halide_current_time_ns());
         #endif
 
 
@@ -222,7 +223,8 @@ static cl_kernel __get_kernel(const char* entry_name) {
     // char msg[1];
     #else
     char msg[256];
-    snprintf(msg, 256, "get_kernel %s (t=%lld)", entry_name, halide_current_time_ns() );
+    snprintf(msg, 256, "get_kernel %s (t=%lld)",
+             entry_name, (long long)halide_current_time_ns() );
     #endif
     // Get kernel function ptr
     TIME_START();
@@ -239,7 +241,8 @@ static cl_mem __dev_malloc(size_t bytes) {
     // char msg[1];
     #else
     char msg[256];
-    snprintf(msg, 256, "dev_malloc (%zu bytes) (t=%lld)", bytes, halide_current_time_ns() );
+    snprintf(msg, 256, "dev_malloc (%zu bytes) (t=%lld)",
+             bytes, (long long)halide_current_time_ns() );
     #endif
     TIME_START();
     int err;
@@ -284,7 +287,8 @@ WEAK void halide_copy_to_dev(buffer_t* buf) {
         // char msg[1];
         #else
         char msg[256];
-        snprintf(msg, 256, "copy_to_dev (%zu bytes) %p -> %p (t=%lld)", size, buf->host, (void*)buf->dev, halide_current_time_ns() );
+        snprintf(msg, 256, "copy_to_dev (%zu bytes) %p -> %p (t=%lld)",
+                 size, buf->host, (void*)buf->dev, (long long)halide_current_time_ns() );
         #endif
         halide_assert(halide_validate_dev_pointer(buf));
         TIME_START();
@@ -334,7 +338,7 @@ WEAK void halide_dev_run(
         msg, 256,
         "dev_run %s with (%dx%dx%d) blks, (%dx%dx%d) threads, %d shmem (t=%lld)",
         entry_name, blocksX, blocksY, blocksZ, threadsX, threadsY, threadsZ, shared_mem_bytes,
-        halide_current_time_ns()
+        (long long)halide_current_time_ns()
     );
     #endif
     // Pack dims
